[step2] refactor: Image 파일 크기 및 파일 길이 클래스 추출
